Add .liftoffrc configuration
This lets users create local or global `.liftoffrc` configuration files to
control the settings used for liftoff.

By default, liftoff will use the internal default configuration, but it
will first try to find a local `.liftoffrc`, or if it can't find that,
then it will try to read the configuration from `~/.liftoffrc`

This works on a key by key basis. If a key isn't defined in a config
file, it will fall back to the next configuration.