Remove command line options
With the addition of a liftoffrc, these command line options lose most
of their value. Their only benefit at this point is the ability to run
liftoff for activating individual settings, which isn't really in line
with the current direction of the tool. I'd rather see people use local
or global liftoffrc files to control these settings.