Remove native C extension step from installation
Xcode 5.1 _seriously_ breaks building gems with the system ruby. And as
it turns out, we don't actually need this in order for Liftoff to work.

It would be great to add this back in later, but in order for that to
happen, we would need to depend on Ruby being updated, or xcodeproj
would need to change the way this is built. Neither of these seem super
likely.