Reluctantly allow the use of tabs for indentation
Everyone knows that you _should_ be using spaces for indentation and not
tabs, but some people insist on being wrong, so we shouldn't make it
hard for them to be wrong.

This adds a `use_tabs` key to the configuration. By default, this key is
false, the way God intended, but people that feel like being insane can
change it to configure their project to use tabs instead of spaces.

Note that this _doesn't_ affect the templates. It's recommended that if
you want to use tabs everywhere, you override the templates to match.