Add the ability to specify arbitrary templates
This adds the ability to define a `templates` key in your `liftoffrc`.
The value for this key should be an array of key/value pairs in the
format of:

<template name>: <destination relative to project root>

We'll add a real default for this key soon, but I'd just like to make
sure this isn't nil for now.