Share created application scheme
In order to share the schemes, we need to do 2 things:

1. We need to create the schemes. This is normally done for us by Xcode
when we open the project, but we can't wait that long.
2. We need to move the application target's generated scheme into the
proper location on disk.

This unfortunately creates a scheme for the unit test target as well,
which is different from Xcode's behavior, but that can't really be
helped right now.