Generate application scheme manually
Letting Xcodeproj create the schemes for us didn't work super well. It
created a scheme for each target individually, and there was no way to
set the test target. We also had to manually move the scheme from the
user's data directory to the shared directory.

Note that at this moment, Xcodeproj sets the testable bundle name to
name.ocunit instead of xctest. Xcode corrects this when it opens the
project, leading to a dirty repo.