Auto-update version and bundle number using Git
This script updates the version info for the build application by
modifying the plist in the biuld binary. It also updates the plist for
the generated dSYM to ensure that they match.

The CFBundleVersion is set based on the number of commits on master
The CFBundleShortVersionString is set based on the most recent git tag