Let users add arbitrary configuration to tests
This adds the `extra_test_config` key to the `liftoffrc`. This key acts
similarly to the `extra_config` key, but adds arbitrary configuration to
the test target. If more than one test target exists, the user will be
given the chance to select the target they would like to modify.