Unset INSTALL_PATH for projects
The default value here was causing really weird issues when trying to
archive apps. It would cause archives to register as Generic Archives
instead of as application archives. Users would either have to edit the
archive directory structure themselves, or delete this setting. We
should just remove it ahead of time, which is consistent with Xcode's
default behavior.