Add support for bundler dependency manager
Bundler was already around in the codebase with a Gemfile, but now it can explicitly included or excluded (default excluded). It copies the Gemfile and runs `bundle install`. It is enabled with the command line flag `--dependency-managers bundler` or in .liftoffrc: `dependency_managers: bundler`

Bundler comes first in the setup script because cocoapods can be installed via bundler.

Support bundler through the rest of the codebase

Shorten options line length

Format liftoffrc manpage

Reorder bundler in script

Bundler is used to install cocoapods so it should come first